requirements.txt
Flask==3.1.2
google-generativeai==0.8.5
python-dotenv==1.1.1
gunicorn==22.0.0
psutil==5.9.8 